moved SNIaLightCurveGenerator to its own file
